feat!: Rename the new serialize module to default serialize
